move LogModal to more generic GameLog
