+++ src/components/game/GameLog/LogStyles.tsx
